Moved files to analyzer package, added Unit enumerator to determine distance scale factor per applied point (default METER).
Added some methods to IComputationPolicy to set important factors for computation.
Added basic implementation of ICommandUsageAnalyzer.
