+++ src/main/java/edu/kis/powp/jobs2d/command/analyzer/Unit.java
@@ -1,0 +1,11 @@
+package edu.kis.powp.jobs2d.command.analyzer;
+
+public enum Unit {
+    NANOMETER,
+    MICROMETER,
+    MILLIMETER,
+    CENTIMETER,
+    DECIMETER,
+    METER,
+    KILOMETER,
+}
