+++ src/main/java/edu/kis/powp/jobs2d/command/analyzer/UsageType.java
@@ -1,0 +1,5 @@
+package edu.kis.powp.jobs2d.command.analyzer;
+
+public enum UsageType {
+    READ,WRITE
+}
