--- src/main/java/edu/kis/powp/jobs2d/command/manager/UsageType.java
@@ -1,5 +1,0 @@
-package edu.kis.powp.jobs2d.command.manager;
-
-public enum UsageType {
-    READ,WRITE
-}
